requirements.txt
certifi==2022.5.18.1
click==8.1.3
colorama==0.4.4
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
mysql-connector==2.2.9
numpy==1.22.4
pandas==1.4.2
pickleshare==0.7.5
pip==21.2.2
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.1.1
scipy==1.8.1
setuptools==61.2.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
Werkzeug==2.1.2
wheel==0.37.1
wincertstore==0.2
zipp==3.8.0
